Refactor app webpack configurations
