--- config/webpack/loaders/index.js
@@ -1,4 +1,0 @@
-module.exports.devLoaders = require('./devLoaders');
-module.exports.prodLoaders = require('./prodLoaders');
-module.exports.testLoaders = require('./testLoaders');
-module.exports.coverageLoaders = require('./coverageLoaders');
