+++ config/webpack/test/webpack.config.base.js
@@ -1,22 +1,17 @@
 const webpack = require('webpack');
+const merge = require('webpack-merge');
+const common = require('../webpack.config.common');
 const helpers = require('../../helpers');
 
-module.exports = {
+module.exports = merge(common, {
   devtool: 'inline-source-map',
   resolve: {
-    extensions: ['.js', '.ts', '.tsx'],
     alias: {
-      globalStyles: helpers.root("src/content/sass"),
       sinon: 'sinon/pkg/sinon',
     },
   },
   module: {
     rules: [
-      {
-        test: /\.tsx?$/,
-        enforce: 'pre',
-        loader: 'tslint-loader',
-      },
       {
         test: /\.css$/,
         include: /node_modules/,
@@ -41,26 +36,6 @@ module.exports = {
           { loader: 'sass-loader' },
         ],
       },
-      {
-        test: /\.(woff|woff2)(\?v=\d+\.\d+\.\d+)?$/,
-        loader: 'url-loader?limit=10000&mimetype=application/font-woff',
-      },
-      {
-        test: /\.ttf(\?v=\d+\.\d+\.\d+)?$/,
-        loader: 'url-loader?limit=10000&mimetype=application/octet-stream',
-      },
-      {
-        test: /\.eot(\?v=\d+\.\d+\.\d+)?$/,
-        loader: 'file-loader',
-      },
-      {
-        test: /\.svg(\?v=\d+\.\d+\.\d+)?$/,
-        loader: 'url-loader?limit=10000&mimetype=image/svg+xml',
-      },
-      {
-        test: /\.(png|jpg|ico)?$/,
-        loader: 'url-loader?limit=10000&mimetype=image/png',
-      },
     ],
     noParse: [
       /node_modules(\\|\/)sinon/,
@@ -71,4 +46,4 @@ module.exports = {
     'react/lib/ExecutionEnvironment': true,
     'react/lib/ReactContext': 'window',
   },
-};
+});
