+++ config/webpack/webpack.config.common.js
@@ -1,0 +1,40 @@
+const webpack = require('webpack');
+const helpers = require('../helpers');
+
+module.exports = {
+  resolve: {
+    extensions: ['.js', '.ts', '.tsx'],
+    alias: {
+      'globalStyles': helpers.root("src/content/sass"),
+    },
+  },
+  module: {
+    rules: [
+      {
+        test: /\.tsx?$/,
+        enforce: 'pre',
+        loader: 'tslint-loader',
+      },
+      {
+        test: /\.(woff|woff2)(\?v=\d+\.\d+\.\d+)?$/,
+        loader: 'url-loader?limit=10000&mimetype=application/font-woff',
+      },
+      {
+        test: /\.ttf(\?v=\d+\.\d+\.\d+)?$/,
+        loader: 'url-loader?limit=10000&mimetype=application/octet-stream',
+      },
+      {
+        test: /\.eot(\?v=\d+\.\d+\.\d+)?$/,
+        loader: 'file-loader',
+      },
+      {
+        test: /\.svg(\?v=\d+\.\d+\.\d+)?$/,
+        loader: 'url-loader?limit=10000&mimetype=image/svg+xml',
+      },
+      {
+        test: /\.(png|jpg|ico)?$/,
+        loader: 'url-loader?limit=10000&mimetype=image/png',
+      },
+    ]
+  }
+}
