+++ package.json
@@ -6,14 +6,14 @@
   "scripts": {
     "postinstall": "npm run build:prod",
     "start": "nf start",
-    "start:dev": "webpack-dev-server --config ./config/webpack/webpack.config.dev.js",
+    "start:dev": "webpack-dev-server --config ./config/webpack/app/webpack.config.dev.js",
     "start:prod": "node server",
     "lint": "tslint -c ./tslint.json --force src/**/*.ts src/**/*.tsx",
     "test": "karma start ./config/karma/karma.conf.test.js",
     "test:watch": "karma start ./config/karma/karma.conf.test.js --no-single-run",
     "test:coverage": "node --max_old_space_size=4096 node_modules/karma/bin/karma start ./config/karma/karma.conf.coverage.js",
-    "build:dev": "rimraf dist && webpack --config ./config/webpack/webpack.config.dev.js",
-    "build:prod": "rimraf public && cross-env NODE_ENV=production webpack --config ./config/webpack/webpack.config.prod.js -p",
+    "build:dev": "rimraf dist && webpack --config ./config/webpack/app/webpack.config.dev.js",
+    "build:prod": "rimraf public && cross-env NODE_ENV=production webpack -p --config ./config/webpack/app/webpack.config.prod.js",
     "clean": "rimraf public"
   },
   "repository": {
